requirements.txt
pillow
mimetypes
